requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2-binary==2.8.6
celery[redis]==5.2.7
flower==1.2.0
celery_singleton==0.3.1
django-redis==5.2.0
django-cachalot==2.5.2